requirements.txt
llama-index-core==0.12.24
llama-index-embeddings-openai==0.3.1
llama-index-llms-openai==0.3.25
llama-index-utils-workflow==0.3.0
llama-parse==0.6.2
nest-asyncio==1.6.0
python-dotenv==1.0.1
